bordercore/aws/create_thumbnail/requirements.txt
boto3==1.9.156
pdf2image==0.1.14
Pillow==6.2.0
PyPDF2==1.26.0
